SDK Check-in: 1.0.4402212019
Update the Language SDK's to reflect the 1.0.4402212019 API spec

+++ libs/java/src/main/java/com/bluejeans/api/rest/meetings/model/Meeting.java
@@ -1,8 +1,8 @@
 /*
  * BlueJeans Meetings REST API
- *  ## Video That Works Where You Do. This site provides developers access to APIs from BlueJean's Meetings meeting service.  From here you will be able to make actual API calls to manage User Accounts, Meetings, and Recordings.  Also, you can pull analytical data and current state information. With these APIs you can quickly integrate **BlueJeans** video administration into your applications.    <alert>7/23/2018 </alert>Corrected errors in API definition file.<br/> 7/10/2018 Deprecated some Command Center API's.  Exposed API's that Summarize usage.<br/> 7/06/2018 Added API for changing ownership of recordings.<br/> 6/28/2018 Restored Create Enterprise Account, added Enterprise and User Profile Tag API's.<br/> 6/01/2018 Clarified Access Token requirements for meeting endpoint API's.<br/> 5/21/2018 Corrected return model for meeting history call.<br/> <hr>  <h2 class=\"info\"> What is a BlueJeans Meeting?</h2> <img src=\"./img/bjnmeeting.png\" style=\"padding-left:20px; width:50%\">  A BlueJeans meeting is a collaboration session of 2 or more participants joining a virtual meeting-room in the cloud.   The first person to join is put into a waiting-room until other participant(s) join.  When the meeting starts, all participants will be connected over video & audio.  <u>Types of Meeting Rooms</u> There are two types of Meeting Rooms available to a registered user.  They are described as: <ul> <li>Scheduled – A room created for a specific date and time.  It is assigned a randomized and unique numeric ID by BlueJeans.</li> <li>Personal – Every user is given a personal meeting room with a customized Meeting Identifier.  People typically use their mobile phone number as their Meeting ID.</li> </ul> <u>Meeting Room Security</u> A meeting room can <i>optionally</i> be protected by a randomized access code.  THe access code ensures security of participants who join.  Also the access code can specify a Moderator role whose user(s) have additional capabilities to administer a meeting. <ul> <li>When set, participants would be required to enter the access code in order to join.</li> <li>The scheduler's profile has a user-specified moderator passcode which can be input to allow designated user(s) greater privileges to help organize meetings.</li> </ul> <hr>  <h2 id='gs' class=\"info\" onclick=\"hideit('gs')\"><span id=\"gsicon\" class=\"glyphicon glyphicon-chevron-right\"></span> Getting Started</h2> <div id=\"gsbody\" style=\"display:none\"> Before you start using BlueJeans' APIs, you must first have a BlueJeans account enabled for API Access.  Contact <a href=\"mailto:Support@BlueJeans.com)\">BlueJeans Support</a> for assistance.  <br /><br />Once you have an account, you may start writing application code to authenticate and make API calls.  *Alternatively*, you can use this developer site to test the BlueJeans' APIs and develop a level of familiarity before you write production code.  <br />  ### To Make API Calls from This Site If you want to use this developer site to try various BlueJeans' APIs, here are the steps required to authenticate and enable your Developer Session to place API calls. 1. Choose Method for Authenticating      * Click on the desired Authentication method from below.      * Click on the **Try It Out** button. 1. Make Authentication request      * Follow APIs instructions and input the API parameters.      * Click on the blue **Execute** button.      * If successful, the API returns with JSON data containing a field called **access_token**.  Copy/save this value. 1. Authorize BlueJeans Developer Session.      * Click on the green **Authorize button**.       * The site will show you a pop-up window for authorization.      * Enter your access token in the field named **api_key**      * Click on the **Authorize** button  Your current BlueJeans developer session is now authenticated and ready to place API calls.  The web site will automatically include your access token on any API calls you make. </div> <hr>  <h2 id='oauth' class=\"info\" onclick=\"hideit('oauth')\"><span id=\"oauthicon\" class=\"glyphicon glyphicon-chevron-right\"></span> About Meetings Authentication</h2>  <div id=\"oauthbody\" style=\"display:none\">  All API transactions (excluding Authentication) require an access token per **OAuth standards**.  BlueJeans provides multiple methods for obtaining an access token.  Additionally there are diffferent scopes of token access. ### Grant Types Bluejeans provides 4 different methods for users to Authenticate.  Successful authentication allows BlueJeans to grant an access token to perform API operations. * Password Credentials Grant – Authenticate with a username and password and receive an access token with user level permission. Known as two-legged OAuth. * Meeting Credentials Grant – Authenticate with a meeting ID and meeting passcode and receive an access token with meeting level permission. Known as two-legged OAuth. * Client Credentials Grant –  Authenticate with a client ID and client secret and receive an access token with enterprise level permission. Known as two-legged OAuth. * Authorization Code Grant – Authentication for your developer's application occurs through a redirection to a BlueJeans authentication page. The application receives an authorization code to be submitted, along with other tokens, to receive an access token. Known as three-legged OAuth. For more information please refer to the [OAuth specification](https://oauth.net/). ### Access & Permissions BlueJeans defines 3 levels of API access into the system.  When an access token is granted, it carries one of these 3 levels.  The scope of system functionality depends upon the token's access level. * Meeting-level – Scope of APIs is limited to individual meetings. * User-level – Scope depends on the requested permissions. * App-level – provisioned either by BlueJeans personnel, or the BlueJeans Enterprise Admin, an app, is issued a client key and secret key. These tokens then are used by the BlueJeans Authentication API to receive the token. The token's scope provides access to the entire enterprise and all of its users. All endpoints in this document that require **Enterprise Admin** access will be marked as such. </div> <hr> 
+ *  ## Video That Works Where You Do. <p>This site provides developers access to APIs from BlueJean's Meetings meeting service.  From here you can see and try actual API calls to manage User Accounts, Meetings, and Recordings, pull Analytical Data and Current State information.</p>    <nosdk> <hr> <div> <div class=\"key\">     <a href=\"./basics.html\" target=\"_\"><img class=\"keyimg\" src=\"./img/bjnmeeting.png\"/></a> </div> <div class=\"keybox\"> Basic instructions about What is a BlueJeans Meeting, and Getting Started Using API's, are <a href=\"./basics.html\" target=\"_blank\">available here <i class=\"glyphicon glyphicon-new-window\"></i></a>. </div> </div> <hr> <alert>2/21/2019</alert> Change response of Change layout and Endpoint API's to 204<br/> 12/03/2018 Add API for setting user profile photograph.<br/> 10/17/2018 Updated API field definitions for time variables to be 64-bit .<br/> 7/23/2018 Corrected errors in API definition file.<br/> 7/10/2018 Deprecated some Command Center API's.  Exposed API's that Summarize usage.<br/> 7/06/2018 Added API for changing ownership of recordings.<br/> 6/28/2018 Restored Create Enterprise Account, added Enterprise and User Profile Tag API's.<br/> 6/01/2018 Clarified Access Token requirements for meeting endpoint API's.<br/> 5/21/2018 Corrected return model for meeting history call.<br/> </nosdk> <hr> 
  *
- * OpenAPI spec version: 1.0.4407232018
+ * OpenAPI spec version: 1.0.4402212019
  * Contact: glenn@bluejeans.com
  *
  * NOTE: This class is auto generated by the swagger code generator program.
@@ -50,10 +50,10 @@ public class Meeting {
   private String description = null;
 
   @SerializedName("start")
-  private Integer start = null;
+  private Long start = null;
 
   @SerializedName("end")
-  private Integer end = null;
+  private Long end = null;
 
   @SerializedName("timezone")
   private String timezone = "America/New_York";
@@ -110,10 +110,10 @@ public class Meeting {
   private Boolean isLargeMeeting = null;
 
   @SerializedName("created")
-  private Integer created = null;
+  private Long created = null;
 
   @SerializedName("lastModified")
-  private Integer lastModified = null;
+  private Long lastModified = null;
 
   @SerializedName("isExpired")
   private Boolean isExpired = null;
@@ -146,10 +146,10 @@ public class Meeting {
   private StartEnd next = null;
 
   @SerializedName("nextStart")
-  private Integer nextStart = null;
+  private Long nextStart = null;
 
   @SerializedName("nextEnd")
-  private Integer nextEnd = null;
+  private Long nextEnd = null;
 
   @SerializedName("isPersonalMeeting")
   private Boolean isPersonalMeeting = false;
@@ -220,7 +220,7 @@ public class Meeting {
     this.description = description;
   }
 
-  public Meeting start(Integer start) {
+  public Meeting start(Long start) {
     this.start = start;
     return this;
   }
@@ -230,15 +230,15 @@ public class Meeting {
    * @return start
   **/
   @ApiModelProperty(required = true, value = "A [UNIX Timestamp](https://currentmillis.com/) in milliseconds")
-  public Integer getStart() {
+  public Long getStart() {
     return start;
   }
 
-  public void setStart(Integer start) {
+  public void setStart(Long start) {
     this.start = start;
   }
 
-  public Meeting end(Integer end) {
+  public Meeting end(Long end) {
     this.end = end;
     return this;
   }
@@ -248,11 +248,11 @@ public class Meeting {
    * @return end
   **/
   @ApiModelProperty(required = true, value = "A [UNIX Timestamp](https://currentmillis.com/) in milliseconds")
-  public Integer getEnd() {
+  public Long getEnd() {
     return end;
   }
 
-  public void setEnd(Integer end) {
+  public void setEnd(Long end) {
     this.end = end;
   }
 
@@ -570,7 +570,7 @@ public class Meeting {
     this.isLargeMeeting = isLargeMeeting;
   }
 
-  public Meeting created(Integer created) {
+  public Meeting created(Long created) {
     this.created = created;
     return this;
   }
@@ -580,15 +580,15 @@ public class Meeting {
    * @return created
   **/
   @ApiModelProperty(value = "This is the epoch-based time (in milliseconds) when the meeting was initially created.")
-  public Integer getCreated() {
+  public Long getCreated() {
     return created;
   }
 
-  public void setCreated(Integer created) {
+  public void setCreated(Long created) {
     this.created = created;
   }
 
-  public Meeting lastModified(Integer lastModified) {
+  public Meeting lastModified(Long lastModified) {
     this.lastModified = lastModified;
     return this;
   }
@@ -598,11 +598,11 @@ public class Meeting {
    * @return lastModified
   **/
   @ApiModelProperty(value = "This is the epoch-based time (in milliseconds) when the meeting was last changed.")
-  public Integer getLastModified() {
+  public Long getLastModified() {
     return lastModified;
   }
 
-  public void setLastModified(Integer lastModified) {
+  public void setLastModified(Long lastModified) {
     this.lastModified = lastModified;
   }
 
@@ -786,7 +786,7 @@ public class Meeting {
     this.next = next;
   }
 
-  public Meeting nextStart(Integer nextStart) {
+  public Meeting nextStart(Long nextStart) {
     this.nextStart = nextStart;
     return this;
   }
@@ -796,15 +796,15 @@ public class Meeting {
    * @return nextStart
   **/
   @ApiModelProperty(value = "")
-  public Integer getNextStart() {
+  public Long getNextStart() {
     return nextStart;
   }
 
-  public void setNextStart(Integer nextStart) {
+  public void setNextStart(Long nextStart) {
     this.nextStart = nextStart;
   }
 
-  public Meeting nextEnd(Integer nextEnd) {
+  public Meeting nextEnd(Long nextEnd) {
     this.nextEnd = nextEnd;
     return this;
   }
@@ -814,11 +814,11 @@ public class Meeting {
    * @return nextEnd
   **/
   @ApiModelProperty(value = "")
-  public Integer getNextEnd() {
+  public Long getNextEnd() {
     return nextEnd;
   }
 
-  public void setNextEnd(Integer nextEnd) {
+  public void setNextEnd(Long nextEnd) {
     this.nextEnd = nextEnd;
   }
